refactor(objects): split the two test files review found still near the cap
Codex review flagged the headroom inventory as incomplete, and it was right in
the way that matters: Batch 2 is entirely Event Enrollment work, and
crates/bacnet-objects/src/event_enrollment/tests.rs sat at 668/700 — 32 lines
of headroom for a batch that will add far more than that. value_types/tests.rs
was at 669.

  event_enrollment/tests.rs  668 -> enrollment 362, fault_parameters 184, alert 124
  value_types/tests.rs       669 -> temporal 298, numeric 136, common 126, strings 115

Split by subject rather than by line count, so a future test lands in an
obvious file.

Largest file in bacnet-objects: 700 -> 652.

Leaf test-function set verified identical before and after (2236 entries, no
addition, removal, or rename); only module paths change.
