+++ crates/bacnet-objects/src/event_enrollment/tests/enrollment.rs
@@ -1,4 +1,8 @@
-use super::*;
+//! EventEnrollmentObject property and lifecycle tests.
+//!
+//! Split out to keep every file under the 700-LOC cap.
+
+use super::super::*;
 
 #[test]
 fn create_event_enrollment() {
@@ -410,341 +414,3 @@ fn write_unknown_property_denied() {
 }
 
 // -----------------------------------------------------------------------
-// FaultParameters tests
-// -----------------------------------------------------------------------
-
-#[test]
-fn fault_parameters_default_none() {
-    let ee = EventEnrollmentObject::new(1, "EE-FP", 0).unwrap();
-    let val = ee
-        .read_property(PropertyIdentifier::FAULT_PARAMETERS, None)
-        .unwrap();
-    assert_eq!(val, PropertyValue::Null);
-}
-
-#[test]
-fn fault_parameters_none_variant() {
-    let mut ee = EventEnrollmentObject::new(1, "EE-FP", 0).unwrap();
-    let fp = FaultParameters::FaultNone;
-    ee.set_fault_parameters(Some(fp.clone()));
-    let val = ee
-        .read_property(PropertyIdentifier::FAULT_PARAMETERS, None)
-        .unwrap();
-    assert_eq!(FaultParameters::decode_property_value(&val).unwrap(), fp);
-}
-
-#[test]
-fn fault_parameters_character_string() {
-    let mut ee = EventEnrollmentObject::new(1, "EE-FP", 0).unwrap();
-    let fp = FaultParameters::FaultCharacterString {
-        fault_values: vec!["alarm".to_string(), "critical".to_string()],
-    };
-    ee.set_fault_parameters(Some(fp.clone()));
-    let val = ee
-        .read_property(PropertyIdentifier::FAULT_PARAMETERS, None)
-        .unwrap();
-    assert_eq!(FaultParameters::decode_property_value(&val).unwrap(), fp);
-}
-
-#[test]
-fn fault_parameters_extended() {
-    let mut ee = EventEnrollmentObject::new(1, "EE-FP", 0).unwrap();
-    let fp = FaultParameters::FaultExtended {
-        vendor_id: 42,
-        extended_fault_type: 7,
-        parameters: vec![0x01, 0x02],
-    };
-    ee.set_fault_parameters(Some(fp.clone()));
-    let val = ee
-        .read_property(PropertyIdentifier::FAULT_PARAMETERS, None)
-        .unwrap();
-    assert_eq!(FaultParameters::decode_property_value(&val).unwrap(), fp);
-}
-
-#[test]
-fn fault_parameters_life_safety() {
-    let mut ee = EventEnrollmentObject::new(1, "EE-FP", 0).unwrap();
-    let ai_oid = ObjectIdentifier::new(ObjectType::ANALOG_INPUT, 1).unwrap();
-    let fp = FaultParameters::FaultLifeSafety {
-        fault_values: vec![1, 2, 3],
-        mode_for_reference: BACnetDeviceObjectPropertyReference {
-            object_identifier: ai_oid,
-            property_identifier: PropertyIdentifier::PRESENT_VALUE.to_raw(),
-            property_array_index: None,
-            device_identifier: None,
-        },
-    };
-    ee.set_fault_parameters(Some(fp.clone()));
-    let val = ee
-        .read_property(PropertyIdentifier::FAULT_PARAMETERS, None)
-        .unwrap();
-    assert_eq!(FaultParameters::decode_property_value(&val).unwrap(), fp);
-}
-
-#[test]
-fn fault_parameters_state() {
-    use bacnet_types::constructed::BACnetPropertyStates;
-    let mut ee = EventEnrollmentObject::new(1, "EE-FP", 0).unwrap();
-    let fp = FaultParameters::FaultState {
-        fault_values: vec![BACnetPropertyStates::BooleanValue(true)],
-    };
-    ee.set_fault_parameters(Some(fp.clone()));
-    let val = ee
-        .read_property(PropertyIdentifier::FAULT_PARAMETERS, None)
-        .unwrap();
-    assert_eq!(FaultParameters::decode_property_value(&val).unwrap(), fp);
-}
-
-#[test]
-fn fault_parameters_status_flags() {
-    let mut ee = EventEnrollmentObject::new(1, "EE-FP", 0).unwrap();
-    let ai_oid = ObjectIdentifier::new(ObjectType::ANALOG_INPUT, 1).unwrap();
-    let fp = FaultParameters::FaultStatusFlags {
-        reference: BACnetDeviceObjectPropertyReference {
-            object_identifier: ai_oid,
-            property_identifier: PropertyIdentifier::STATUS_FLAGS.to_raw(),
-            property_array_index: None,
-            device_identifier: None,
-        },
-    };
-    ee.set_fault_parameters(Some(fp.clone()));
-    let val = ee
-        .read_property(PropertyIdentifier::FAULT_PARAMETERS, None)
-        .unwrap();
-    assert_eq!(FaultParameters::decode_property_value(&val).unwrap(), fp);
-}
-
-#[test]
-fn fault_parameters_out_of_range() {
-    let mut ee = EventEnrollmentObject::new(1, "EE-FP", 0).unwrap();
-    let fp = FaultParameters::FaultOutOfRange {
-        min_normal: 0.0,
-        max_normal: 100.0,
-    };
-    ee.set_fault_parameters(Some(fp.clone()));
-    let val = ee
-        .read_property(PropertyIdentifier::FAULT_PARAMETERS, None)
-        .unwrap();
-    assert_eq!(FaultParameters::decode_property_value(&val).unwrap(), fp);
-}
-
-#[test]
-fn fault_parameters_listed() {
-    let mut ee = EventEnrollmentObject::new(1, "EE-FP", 0).unwrap();
-    let ai_oid = ObjectIdentifier::new(ObjectType::ANALOG_INPUT, 1).unwrap();
-    let fp = FaultParameters::FaultListed {
-        reference: BACnetDeviceObjectPropertyReference {
-            object_identifier: ai_oid,
-            property_identifier: PropertyIdentifier::PRESENT_VALUE.to_raw(),
-            property_array_index: None,
-            device_identifier: None,
-        },
-    };
-    ee.set_fault_parameters(Some(fp.clone()));
-    let val = ee
-        .read_property(PropertyIdentifier::FAULT_PARAMETERS, None)
-        .unwrap();
-    assert_eq!(FaultParameters::decode_property_value(&val).unwrap(), fp);
-}
-
-#[test]
-fn fault_parameters_in_property_list() {
-    let ee = EventEnrollmentObject::new(1, "EE-FP", 0).unwrap();
-    let props = ee.property_list();
-    assert!(props.contains(&PropertyIdentifier::FAULT_PARAMETERS));
-}
-
-#[test]
-fn fault_parameters_write_round_trip() {
-    let mut ee = EventEnrollmentObject::new(1, "EE-FP", 0).unwrap();
-    let fp = FaultParameters::FaultOutOfRange {
-        min_normal: -5.0,
-        max_normal: 55.0,
-    };
-    ee.write_property(
-        PropertyIdentifier::FAULT_PARAMETERS,
-        None,
-        fp.encode_property_value(),
-        None,
-    )
-    .unwrap();
-    let val = ee
-        .read_property(PropertyIdentifier::FAULT_PARAMETERS, None)
-        .unwrap();
-    assert_eq!(FaultParameters::decode_property_value(&val).unwrap(), fp);
-}
-
-#[test]
-fn fault_parameters_write_clear_to_null() {
-    let mut ee = EventEnrollmentObject::new(1, "EE-FP", 0).unwrap();
-    ee.set_fault_parameters(Some(FaultParameters::FaultNone));
-    ee.write_property(
-        PropertyIdentifier::FAULT_PARAMETERS,
-        None,
-        PropertyValue::Null,
-        None,
-    )
-    .unwrap();
-    let val = ee
-        .read_property(PropertyIdentifier::FAULT_PARAMETERS, None)
-        .unwrap();
-    assert_eq!(val, PropertyValue::Null);
-}
-
-#[test]
-fn fault_parameters_clear() {
-    let mut ee = EventEnrollmentObject::new(1, "EE-FP", 0).unwrap();
-    ee.set_fault_parameters(Some(FaultParameters::FaultNone));
-    let val = ee
-        .read_property(PropertyIdentifier::FAULT_PARAMETERS, None)
-        .unwrap();
-    assert_eq!(
-        FaultParameters::decode_property_value(&val).unwrap(),
-        FaultParameters::FaultNone
-    );
-
-    // Clear back to None
-    ee.set_fault_parameters(None);
-    let val = ee
-        .read_property(PropertyIdentifier::FAULT_PARAMETERS, None)
-        .unwrap();
-    assert_eq!(val, PropertyValue::Null);
-}
-
-// -----------------------------------------------------------------------
-// AlertEnrollmentObject tests
-// -----------------------------------------------------------------------
-
-#[test]
-fn alert_enrollment_create() {
-    let ae = AlertEnrollmentObject::new(1, "AE-1").unwrap();
-    assert_eq!(
-        ae.object_identifier().object_type(),
-        ObjectType::ALERT_ENROLLMENT
-    );
-    assert_eq!(ae.object_identifier().instance_number(), 1);
-    assert_eq!(ae.object_name(), "AE-1");
-}
-
-#[test]
-fn alert_enrollment_object_type() {
-    let ae = AlertEnrollmentObject::new(1, "AE").unwrap();
-    let val = ae
-        .read_property(PropertyIdentifier::OBJECT_TYPE, None)
-        .unwrap();
-    assert_eq!(
-        val,
-        PropertyValue::Enumerated(ObjectType::ALERT_ENROLLMENT.to_raw())
-    );
-}
-
-#[test]
-fn alert_enrollment_present_value_default() {
-    let ae = AlertEnrollmentObject::new(1, "AE").unwrap();
-    let val = ae
-        .read_property(PropertyIdentifier::PRESENT_VALUE, None)
-        .unwrap();
-    assert_eq!(val, PropertyValue::Enumerated(0));
-}
-
-#[test]
-fn alert_enrollment_event_detection_enable() {
-    let ae = AlertEnrollmentObject::new(1, "AE").unwrap();
-    let val = ae
-        .read_property(PropertyIdentifier::EVENT_DETECTION_ENABLE, None)
-        .unwrap();
-    assert_eq!(val, PropertyValue::Boolean(true));
-}
-
-#[test]
-fn alert_enrollment_write_event_detection_enable() {
-    let mut ae = AlertEnrollmentObject::new(1, "AE").unwrap();
-    ae.write_property(
-        PropertyIdentifier::EVENT_DETECTION_ENABLE,
-        None,
-        PropertyValue::Boolean(false),
-        None,
-    )
-    .unwrap();
-    let val = ae
-        .read_property(PropertyIdentifier::EVENT_DETECTION_ENABLE, None)
-        .unwrap();
-    assert_eq!(val, PropertyValue::Boolean(false));
-}
-
-#[test]
-fn alert_enrollment_event_enable() {
-    let ae = AlertEnrollmentObject::new(1, "AE").unwrap();
-    let val = ae
-        .read_property(PropertyIdentifier::EVENT_ENABLE, None)
-        .unwrap();
-    // Default event_enable = 0b111, shifted left 5 = 0b1110_0000
-    assert_eq!(
-        val,
-        PropertyValue::BitString {
-            unused_bits: 5,
-            data: vec![0b1110_0000],
-        }
-    );
-}
-
-#[test]
-fn alert_enrollment_write_event_enable() {
-    let mut ae = AlertEnrollmentObject::new(1, "AE").unwrap();
-    ae.write_property(
-        PropertyIdentifier::EVENT_ENABLE,
-        None,
-        PropertyValue::BitString {
-            unused_bits: 5,
-            data: vec![0b1000_0000],
-        },
-        None,
-    )
-    .unwrap();
-    let val = ae
-        .read_property(PropertyIdentifier::EVENT_ENABLE, None)
-        .unwrap();
-    assert_eq!(
-        val,
-        PropertyValue::BitString {
-            unused_bits: 5,
-            data: vec![0b1000_0000],
-        }
-    );
-}
-
-#[test]
-fn alert_enrollment_notification_class() {
-    let ae = AlertEnrollmentObject::new(1, "AE").unwrap();
-    let val = ae
-        .read_property(PropertyIdentifier::NOTIFICATION_CLASS, None)
-        .unwrap();
-    assert_eq!(val, PropertyValue::Unsigned(0));
-}
-
-#[test]
-fn alert_enrollment_write_notification_class() {
-    let mut ae = AlertEnrollmentObject::new(1, "AE").unwrap();
-    ae.write_property(
-        PropertyIdentifier::NOTIFICATION_CLASS,
-        None,
-        PropertyValue::Unsigned(42),
-        None,
-    )
-    .unwrap();
-    let val = ae
-        .read_property(PropertyIdentifier::NOTIFICATION_CLASS, None)
-        .unwrap();
-    assert_eq!(val, PropertyValue::Unsigned(42));
-}
-
-#[test]
-fn alert_enrollment_property_list() {
-    let ae = AlertEnrollmentObject::new(1, "AE").unwrap();
-    let props = ae.property_list();
-    assert!(props.contains(&PropertyIdentifier::PRESENT_VALUE));
-    assert!(props.contains(&PropertyIdentifier::EVENT_DETECTION_ENABLE));
-    assert!(props.contains(&PropertyIdentifier::EVENT_ENABLE));
-    assert!(props.contains(&PropertyIdentifier::NOTIFICATION_CLASS));
-    assert!(props.contains(&PropertyIdentifier::STATUS_FLAGS));
-}
