+++ crates/bacnet-objects/src/event_enrollment/tests/mod.rs
@@ -1,0 +1,3 @@
+mod alert;
+mod enrollment;
+mod fault_parameters;
