+++ crates/bacnet-objects/src/value_types/tests/mod.rs
@@ -1,0 +1,4 @@
+mod common;
+mod numeric;
+mod strings;
+mod temporal;
